refactor: move cron logic to a utility fn
